requirements.txt
requests
schedule
Flask-APScheduler
openai == 1.30.1
python-dotenv
Flask==3.0.3